requirements.txt
click~=8.0.0
Flask~=2.0.0
itsdangerous~=2.0.0
Jinja2~=3.0.0
MarkupSafe~=2.0.0
pip~=21.1.1
setuptools~=56.2.0
Werkzeug~=2.0.1
boto3~=1.17.74
python-dotenv~=0.17.1
